+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Universal Scene Description ASCII (USDA)
 - Universal Scene Description Binary (USDC)
 - Universal Scene Description Zipped (USDZ)
+- Virtual Reality Modeling Language (VRML)
 
 # Version 0.18.0 (2023-07-14)